feat(heromodel): replaced the 3D Model in the hero section
deleted the old desktop_pc model, which was not made by myself, and replaced it with an macbook
model, which I made. Also adjusted it to look like I want on bigger and mobile screens.
