+++ src/components/canvas/index.js
@@ -1,5 +1,5 @@
 import EarthCanvas from "./Earth";
-import ComputersCanvas from "./Computers";
+import MacbookCanvas from "./Macbook";
 import StarsCanvas from "./Stars";
 
-export { EarthCanvas, ComputersCanvas, StarsCanvas };
+export { EarthCanvas, MacbookCanvas, StarsCanvas };
